pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
replit = "^3.2.4"
"discord.py" = "^1.7.3"
PyNaCl = "^1.5.0"
ffmpeg = "^1.4"
opus-api = "^0.6.2"
opuslib = "^3.0.1"
gTTS = "^2.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"